backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-json-logger<3.0.0
